+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.7.0
+
+- Remove the cruky_cli
+- Adding hotreload instead of auto reload
+- adding `ServerApp` for the server settings
+- adding multi threaded option with isolates in the release mode (Not in the debug mode)
+
 ## 0.6.0
 
 - Changing the library design